+++ deployment/terraform-module-knfsd/main.tf
@@ -9,7 +9,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "~> 6.34.0"
+      version = "~> 6.36.0"
     }
     random = {
       source  = "hashicorp/random"
@@ -22,7 +22,7 @@ terraform {
   }
   provider_meta "aws" {
     user_agent = [
-      "knfsd-file-cache/terraform-module-knfsd/1.1.0-alpha.22"
+      "knfsd-file-cache/terraform-module-knfsd/1.1.0-alpha.23"
     ]
   }
 }