+++ examples/nfs-standard/main.tf
@@ -9,12 +9,12 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "~> 6.34.0"
+      version = "~> 6.36.0"
     }
   }
   provider_meta "aws" {
     user_agent = [
-      "knfsd-file-cache/examples/nfs-standard/1.1.0-alpha.22"
+      "knfsd-file-cache/examples/nfs-standard/1.1.0-alpha.23"
     ]
   }
 }